extended trickyOneHandler messages from a base Message to reduce code smells
