+++ core/src/main/java/com/mankomania/game/core/network/messages/servertoclient/trickyone/EndTrickyOne.java
@@ -1,14 +1,15 @@
 package com.mankomania.game.core.network.messages.servertoclient.trickyone;
 
+import com.mankomania.game.core.network.messages.clienttoserver.trickyone.TrickyOneBaseMessage;
+
 import java.util.Objects;
 
 /*********************************
  Created by Fabian Oraze on 23.05.20
  *********************************/
 
-public class EndTrickyOne {
+public class EndTrickyOne extends TrickyOneBaseMessage {
 
-    private int playerIndex;
     private int amountWinLose;
 
     //only used to inform clients that MiniGame has ended and they can switch back to main Screen
@@ -19,14 +20,6 @@ public class EndTrickyOne {
     public EndTrickyOne(int playerIndex, int amountWinLose) {
         this.playerIndex = playerIndex;
         this.amountWinLose = amountWinLose;
-    }
-
-    public int getPlayerIndex() {
-        return playerIndex;
-    }
-
-    public void setPlayerIndex(int playerIndex) {
-        this.playerIndex = playerIndex;
     }
 
     public int getAmountWinLose() {
